[Table] Use Vue SFCs
Use Vue SFCs.  Use inject/provide to pass services to components
instead of wrapping components in closures.
