+++ src/plugins/telemetryTable/table-row.html
