--- src/plugins/telemetryTable/telemetry-table.html
@@ -1,64 +1,0 @@
-<div class="tabular-holder l-sticky-headers has-control-bar l-telemetry-table" :class="{'loading': loading}">
-    <div class="l-control-bar">
-        <a class="s-button t-export icon-download labeled"
-           v-on:click="exportAsCSV()"
-           title="Export This View's Data">
-            Export As CSV
-        </a>
-    </div>
-    <!-- Headers table -->
-    <div class="mct-table-headers-w">
-        <table class="mct-table l-tabular-headers filterable" :style="{ 'max-width': totalWidth + 'px'}">
-            <thead>
-                <tr>
-                    <th v-for="(title, key, headerIndex) in headers"
-                        v-on:click="sortBy(key)"
-                        :class="['sortable', sortOptions.key === key ? 'sort' : '', sortOptions.direction].join(' ')"
-                        :style="{ width: columnWidths[headerIndex], 'max-width': columnWidths[headerIndex]}">{{title}}</th>
-                </tr>
-                <tr class="s-filters">
-                    <th v-for="(title, key, headerIndex) in headers"
-                        :style="{
-                            width: columnWidths[headerIndex],
-                            'max-width': columnWidths[headerIndex],
-                        }">
-                        <div class="holder l-filter flex-elem grows" :class="{active: filters[key]}">
-                            <input type="text" v-model="filters[key]" v-on:input="filterChanged(key)" />
-                            <a class="clear-icon clear-input icon-x-in-circle" :class="{show: filters[key]}" @click="clearFilter(key)"></a>
-                        </div>
-                    </th>
-                    </tr>
-            </thead>
-        </table>
-    </div>
-    <!-- Content table -->
-    <div @scroll="scroll" class="l-tabular-body t-scrolling vscroll--persist">
-        <div class="mct-table-scroll-forcer"
-            :style="{
-                width: totalWidth
-            }"></div>
-        <table class="mct-table js-telemetry-table" :style="{ height: totalHeight + 'px', 'max-width': totalWidth + 'px'}">
-            <tbody>
-                <telemetry-table-row v-for="(row, rowIndex) in visibleRows" 
-                    :headers="headers"
-                    :columnWidths="columnWidths"
-                    :rowIndex="rowIndex"
-                    :rowOffset="rowOffset"
-                    :rowHeight="rowHeight"
-                    :row="row"
-                    >
-                </telemetry-table-row>
-            </tbody>
-        </table>
-    </div>
-    <!-- Sizing table -->
-    <table class="mct-sizing-table t-sizing-table js-sizing-table" :style="{width: calcTableWidth}">
-        <tr>
-            <th v-for="(title, key, headerIndex) in headers">{{title}}</th>
-        </tr>
-        <telemetry-table-row v-for="(sizingRowData, objectKeyString) in sizingRows"
-            :headers="headers"
-            :row="sizingRowData">
-        </telemetry-table-row>
-    </table>
-</div>
